+++ changelog.md
@@ -4,6 +4,7 @@
 * `stf.bone`: added `tr` & `tr_armature` properties. They store transforms relative to the parent and armature respectively. `translation` and `rotation` are deprecated.
 * `stf.instance.armature`: Deprecated animation path part `component_mods`, replaced with `components`.
 * AVA UNIVRM0 detection uses assembly version define.
+* UNIVRM0 eye rotation limits get correctly converted.
 * Started implementing a VRM1 context.
 
 ## v0.1.4